+++ src/server.tsx
@@ -5,441 +5,14 @@ import { etag } from 'hono/etag'
 import { cors } from 'hono/cors'
 import { serveStatic } from 'hono/bun'
 import config from 'config'
-import NodeCache from 'node-cache';
-import RpcClient from 'bitcoind-rpc'
-import pino, { type Logger } from 'pino'
 
-// Get application configuration values from the config package.
-const PORT = config.get<number>('server.port');
-const BASE_URL = config.get<number>('server.baseUrl');
+import { getEstimates, BASE_URL, CACHE_STDTTL, PORT, LOGLEVEL } from './util';
+import { logger } from './logger';
 
-const ESPLORA_BASE_URL = config.get<string>('esplora.baseUrl');
-const ESPLORA_FALLBACK_BASE_URL = config.get<string>('esplora.fallbackBaseUrl');
-
-const MEMPOOL_BASE_URL = config.get<string>('mempool.baseUrl');
-const MEMPOOL_FALLBACK_BASE_URL = config.get<string>('mempool.fallbackBaseUrl');
-const MEMPOOL_DEPTH = config.get<number>('mempool.depth');
-
-const BITCOIND_BASE_URL = config.get<string>('bitcoind.baseUrl');
-const BITCOIND_USERNAME = config.get<string>('bitcoind.username');
-const BITCOIND_PASSWORD = config.get<number>('bitcoind.password');
-const BITCOIND_CONF_TARGETS = config.get<number[]>('bitcoind.confTargets');
-const BITCOIND_ESTIMATE_MODE = config.get<EstimateMode>('bitcoind.estimateMode');
-
-const LOGLEVEL = config.get<string>('settings.loglevel');
-const TIMEOUT = config.get<number>('settings.timeout');
-const FEE_MULTIPLIER = config.get<number>('settings.feeMultiplier');
-const FEE_MINIMUM = config.get<number>('settings.feeMinimum');
-const CACHE_STDTTL = config.get<number>('cache.stdTTL');
-const CACHE_CHECKPERIOD = config.get<number>('cache.checkperiod');
-
-let logger : Logger;
-const pinoOptions = {
-  level: LOGLEVEL,
-  messageKey: 'message',
-  formatters: {
-    level: (label: string) => {
-      return { level: label };
-    }
-  }
-}
-if (process.env.NODE_ENV === 'production') {
-  logger = pino(pinoOptions);
-}
-else {
-  try {
-    const pretty = require('pino-pretty');
-    logger = pino(pinoOptions, pretty());
-  } catch (error) {
-    logger = pino(pinoOptions);
-  }
-}
+const log = logger(LOGLEVEL);
 
 // Log the configuration values.
-logger.info({
-  mempoolSettings: {
-    baseUrl: MEMPOOL_BASE_URL,
-    fallbackBaseUrl: MEMPOOL_FALLBACK_BASE_URL,
-    estimationDepth: MEMPOOL_DEPTH
-  },
-  esploraSettings: {
-    baseUrl: ESPLORA_BASE_URL,
-    fallbackBaseUrl: ESPLORA_FALLBACK_BASE_URL
-  },
-  bitcoindSettings: {
-    baseUrl: BITCOIND_BASE_URL,
-    username: BITCOIND_USERNAME,
-    password: '******'
-  },
-  appSettings: {
-    serverPort: PORT,
-    baseUrl: BASE_URL,
-    feeMinimum: FEE_MINIMUM,
-    feeMultiplier: FEE_MULTIPLIER,
-    timeout: TIMEOUT,
-    cacheStdTTL: CACHE_STDTTL,
-    cacheCheckPeriod: CACHE_CHECKPERIOD
-  }
-});
-
-// Constants
-const MEMPOOL_TIP_HASH_URL = MEMPOOL_BASE_URL && `${MEMPOOL_BASE_URL}/api/blocks/tip/hash`;
-const ESPLORA_TIP_HASH_URL = ESPLORA_BASE_URL && `${ESPLORA_BASE_URL}/api/blocks/tip/hash`;
-const MEMPOOL_FEES_URL = MEMPOOL_BASE_URL && `${MEMPOOL_BASE_URL}/api/v1/fees/recommended`;
-const ESPLORA_FEE_ESTIMATES_URL = ESPLORA_BASE_URL && `${ESPLORA_BASE_URL}/api/fee-estimates`;
-
-const MEMPOOL_TIP_HASH_URL_FALLBACK = MEMPOOL_FALLBACK_BASE_URL && `${MEMPOOL_FALLBACK_BASE_URL}/api/blocks/tip/hash`;
-const ESPLORA_TIP_HASH_URL_FALLBACK = ESPLORA_FALLBACK_BASE_URL && `${ESPLORA_FALLBACK_BASE_URL}/api/blocks/tip/hash`;
-const MEMPOOL_FEES_URL_FALLBACK = MEMPOOL_FALLBACK_BASE_URL && `${MEMPOOL_FALLBACK_BASE_URL}/api/v1/fees/recommended`;
-const ESPLORA_FEE_ESTIMATES_URL_FALLBACK = ESPLORA_FALLBACK_BASE_URL && `${ESPLORA_FALLBACK_BASE_URL}/api/fee-estimates`;
-
-// Initialize the cache.
-const CACHE_KEY = 'estimates';
-const cache = new NodeCache({ stdTTL: CACHE_STDTTL, checkperiod: CACHE_CHECKPERIOD });
-
-
-/**
- * Helper function to extract value from a fulfilled promise.
- */
-function getValueFromFulfilledPromise(result: PromiseSettledResult<any>) {
-  return result && result.status === "fulfilled" && result.value ? result.value : null;
-}
-
-// NOTE: fetch signal abortcontroller does not work on Bun.
-// See https://github.com/oven-sh/bun/issues/2489
-async function fetchWithTimeout(url: string, timeout: number = TIMEOUT): Promise<Response> {
-  logger.debug({ message: `Starting fetch request to ${url}` });
-  const fetchPromise = fetch(url);
-  const timeoutPromise = new Promise((_, reject) => 
-    setTimeout(() => reject(new Error(`Request timed out after ${timeout} ms`)), timeout)
-  );
-
-  return Promise.race([fetchPromise, timeoutPromise]) as Promise<Response>;
-}
-
-/**
- * Fetches data from the given URL and validates and processes the response.
- */
-async function fetchAndProcess(url: string, expectedResponseType: ExpectedResponseType): Promise<string | object | null> {
-  const response = await fetchWithTimeout(url, TIMEOUT);
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-  logger.debug({ message: `Successfully fetched data from ${url}` });
-
-  const contentType = response.headers.get("content-type");
-  if (expectedResponseType === 'json' && contentType?.includes("application/json")) {
-    return response.json();
-  } else if (expectedResponseType === 'text' && contentType?.includes("text/plain")) {
-    const text = await response.text();
-    const trimmedText = text.trim();
-    if (trimmedText.includes('\n') || text !== trimmedText) {
-      throw new Error('Response is not a single text string with no whitespace or newlines');
-    }
-    return trimmedText;
-  } else {
-    throw new Error(`Unexpected response type. Expected ${expectedResponseType}, but received ${contentType}`);
-  }
-}
-
-/**
- * Fetches data from the given URL with a timeout, fallback, and error handling.
- */
-async function fetchAndHandle(url: string, expectedResponseType: ExpectedResponseType, fallbackUrl?: string): Promise<string | object | null> {
-  try {
-    const timeout = new Promise((resolve) => setTimeout(resolve, TIMEOUT, 'timeout'));
-    const fetchPromise = fetchAndProcess(url, expectedResponseType);
-    const result = await Promise.race([fetchPromise, timeout]) as Promise<Response> | string;
-
-    if (result === 'timeout' || result instanceof Error) {
-      throw new Error('Fetch or timeout error');
-    }
-
-    return result;
-  } catch (error) {
-    if (fallbackUrl) {
-      logger.debug({ message: 'Trying fallback URL: ${fallbackUrl}' });
-      return fetchAndProcess(fallbackUrl, expectedResponseType);
-    } else {
-      throw new Error(`Fetch request to ${url} failed and no fallback URL was provided.`);
-    }
-  }
-}
-
-/**
- * Fetches mempool fees.
- */
-async function fetchMempoolData() : Promise<MempoolFeeEstimates | null> {
-  const tasks = [
-    MEMPOOL_FEES_URL && fetchAndHandle(MEMPOOL_FEES_URL, 'json'),
-    MEMPOOL_FEES_URL_FALLBACK && fetchAndHandle(MEMPOOL_FEES_URL_FALLBACK, 'json'),
-  ].filter(Boolean);
-  if (tasks.length === 0) {
-    return null;
-  }
-
-  const results = await Promise.allSettled(tasks);
-  logger.debug({ message: 'Fetched data from mempool: {results}', results });
-
-  let res0 = getValueFromFulfilledPromise(results[0]);
-  let res1 = getValueFromFulfilledPromise(results[1]);
-
-  // If all of the response properties are 1, then the response is an error (probably the mempool data is not available).
-  const isRes0Invalid = !res0 || (Object.values(res0).every((value) => value === 1));
-  const isRes1Invalid = !res1 || (Object.values(res1).every((value) => value === 1));
-
-  // Return a response that is valid, or null if both responses are invald.
-  let data
-  if (!isRes0Invalid) {
-    data = res0;
-  } else {
-    data = isRes1Invalid ? null : res1;
-  }
-  logger.info({ message: 'Using data from mempool: {data}', data });
-  return data;
-}
-
-/**
- * Fetches esplora fees.
- */
-async function fetchEsploraData() : Promise<FeeByBlockTarget | null> {
-  const tasks = [
-    ESPLORA_FEE_ESTIMATES_URL && fetchAndHandle(ESPLORA_FEE_ESTIMATES_URL, 'json'),
-    ESPLORA_FEE_ESTIMATES_URL_FALLBACK && fetchAndHandle(ESPLORA_FEE_ESTIMATES_URL_FALLBACK, 'json'),
-  ].filter(Boolean);
-  if (tasks.length === 0) {
-    return null;
-  }
-  const results = await Promise.allSettled(tasks);
-  logger.debug({ message: 'Fetched data from esplora: {results}', results });
-
-  let res0 = getValueFromFulfilledPromise(results[0]);
-  let res1 = getValueFromFulfilledPromise(results[1]);
-
-  const data = res0 || res1 || null;
-  logger.info({ message: 'Using data from esplora: {data}', data });
-  return data;
-}
-
-/**
- * Fetches bitcoind fees.
- */
-async function fetchBitcoindData() : Promise<FeeByBlockTarget | null> {
-  if (!BITCOIND_BASE_URL) {
-    return null;
-  }
-
-  return new Promise((resolve, _) => {
-    let data : FeeByBlockTarget = {};
-
-    // Define the targets for which to fetch fee estimates.
-    const targets = BITCOIND_CONF_TARGETS;
-
-    // Extract protocol, host, port from bitcoindBaseUrl.
-    let { protocol, hostname: host, port } = new URL(BITCOIND_BASE_URL);
-
-    // Strip the trailing colon from the protocol.
-    protocol = protocol.replace(/.$/, '')
-
-    const config = {
-      protocol,
-      host,
-      port,
-      user: BITCOIND_USERNAME,
-      pass: BITCOIND_PASSWORD,
-    };
-
-    const rpc = new RpcClient(config);
-
-    function batchCall() {
-      targets.forEach(function (target) {
-        rpc.estimatesmartfee(target, BITCOIND_ESTIMATE_MODE);
-      });
-    }
-
-    rpc.batch(batchCall, (error: Error | null, response: BitcoindRpcBatchResponse[]) => {
-      if (error) {
-        logger.error({ message: 'Unable to fetch fee estimates from bitcoind: {error}', error });
-        resolve(null);
-      } else {
-        targets.forEach((target, i) => {  
-          let feeRate = response[i].result?.feerate;
-          if (feeRate) {
-            // convert the returned value to satoshis, as it's currently returned in BTC.
-            data[target] = feeRate * 1e8;
-          } else {
-            logger.error({ message: `Failed to fetch fee estimate from bitcoind for confirmation target ${target}: {errors}`,
-              errors: response[i].result?.errors});
-          }
-        });
-        logger.info({ message: 'Using data from bitcoind: {data}', data });
-        resolve(data);
-      }
-    });
-  });
-}
-
-function processEstimates(estimates: FeeByBlockTarget, applyMultiplier = true, convert = false) : FeeByBlockTarget {
-  for (const [blockTarget, fee] of Object.entries(estimates) ) {
-    let estimate = fee;
-    if (applyMultiplier) {
-      estimate = estimate * FEE_MULTIPLIER;
-    }
-    if (convert) {
-      estimate = estimate * 1000;
-    }
-    estimates[Number(blockTarget)] = Math.ceil(estimate);
-  }
-  return estimates;
-}
-
-/**
- * Fetches the current block hash.
- */
-async function fetchBlocksTipHash() : Promise<string | null> {
-  const tasks = [
-    (MEMPOOL_TIP_HASH_URL || MEMPOOL_TIP_HASH_URL_FALLBACK) && fetchAndHandle(MEMPOOL_TIP_HASH_URL, 'text', MEMPOOL_TIP_HASH_URL_FALLBACK),
-    (ESPLORA_TIP_HASH_URL || ESPLORA_TIP_HASH_URL_FALLBACK) && fetchAndHandle(ESPLORA_TIP_HASH_URL, 'text', ESPLORA_TIP_HASH_URL_FALLBACK),
-  ].filter(Boolean);
-  const res = await Promise.allSettled(tasks);
-
-  let res0 = getValueFromFulfilledPromise(res[0]);
-  let res1 = getValueFromFulfilledPromise(res[1]);
-
-  return res0 || res1 || null;
-}
-
-/**
- * Gets the current fee estimates from the cache or fetches them if they are not cached.
- */
-async function getEstimates() : Promise<Estimates> {
-  let estimates: Estimates | undefined = cache.get(CACHE_KEY);
-
-  if (estimates) {
-    logger.info({ message: 'Got estimates from cache: ${estimates}', estimates });
-    return estimates;
-  }
-
-  const tasks = [
-    await fetchMempoolData(),
-    await fetchEsploraData(),
-    await fetchBitcoindData(),
-    await fetchBlocksTipHash(),
-  ];
-  const [result1, result2, result3, result4] = await Promise.allSettled(tasks);
-  const mempoolFeeEstimates = getValueFromFulfilledPromise(result1);
-  const esploraFeeEstimates = getValueFromFulfilledPromise(result2);
-  const bitcoindFeeEstimates = getValueFromFulfilledPromise(result3);
-  const blocksTipHash = getValueFromFulfilledPromise(result4);
-
-  estimates = {
-    current_block_hash: blocksTipHash,
-    fee_by_block_target: calculateFees(mempoolFeeEstimates, esploraFeeEstimates, bitcoindFeeEstimates),
-  };
-
-  cache.set(CACHE_KEY, estimates);
-
-  logger.info({ message: 'Got estimates: {estimates}', estimates });
-  return estimates;
-}
-
-/**
- * Get the fee estimates that are above the desired mempool depth.
- */
-function extractMempoolFees(mempoolFeeEstimates: MempoolFeeEstimates): FeeByBlockTarget {
-  const feeByBlockTarget: FeeByBlockTarget = {};
-
-  if (mempoolFeeEstimates) {
-    const blockTargetMapping: BlockTargetMapping = {
-      1: 'fastestFee',
-      3: 'halfHourFee',
-      6: 'hourFee'
-    };
-    for (let i = 1; i <= MEMPOOL_DEPTH; i++) {
-      const feeProperty = blockTargetMapping[i];
-      if (feeProperty && mempoolFeeEstimates[feeProperty]) {
-        feeByBlockTarget[i] = mempoolFeeEstimates[feeProperty];
-      }
-    }
-  }
-
-  return feeByBlockTarget;
-}
-
-/**
- * Gets the lowest fee from the feeByBlockTarget object.
- */
-function getLowestFee(feeByBlockTarget: FeeByBlockTarget) : number | null {
-  const values = Object.values(feeByBlockTarget);
-  return values.length > 0 ? Math.min(...values) : null;
-}
-
-/**
- * Filters the estimates to remove duplicates and estimates that are lower than the desired minimum fee.
- */
-function filterEstimates(feeByBlockTarget: FeeByBlockTarget, minFee: number): FeeByBlockTarget {
-  const result: FeeByBlockTarget = {};
-
-  for (const [blockTarget, fee] of Object.entries(feeByBlockTarget)) {
-    if (fee >= minFee) {
-      result[Number(blockTarget)] = fee;
-    }
-  }
-
-  // If we didn't manage to get any fee estimates, return a single estimate with the minimum fee.
-  if (Object.keys(result).length === 0) {
-    result[1] = minFee;
-  }
-
-  return result;
-}
-
-/**
- * Appends estimates to the feeByBlockTarget object.
- */
-function addFeeEstimates(feeByBlockTarget: FeeByBlockTarget, feeEstimates: FeeByBlockTarget) {
-  const lowestFee = getLowestFee(feeByBlockTarget)
-  for (const [blockTarget, fee] of Object.entries(feeEstimates)) {
-    if (!lowestFee || fee < lowestFee) {
-      feeByBlockTarget[Number(blockTarget)] = Math.ceil(fee);
-    }
-  }
-}
-
-/**
- * Calculates the fees.
- */
-function calculateFees(mempoolFeeEstimates: MempoolFeeEstimates, esploraFeeEstimates: FeeByBlockTarget, bitcoindFeeEstimates: FeeByBlockTarget) {
-  let feeByBlockTarget: FeeByBlockTarget = {};
-
-  // Get the mempool fee estimates.
-  if (mempoolFeeEstimates) {
-    let estimates = extractMempoolFees(mempoolFeeEstimates);
-    estimates = processEstimates(estimates, true, true);
-    addFeeEstimates(feeByBlockTarget, estimates);
-  }
-
-  // Add the bitcoind fee estimates.
-  if (bitcoindFeeEstimates) {
-    const estimates = processEstimates(bitcoindFeeEstimates, true, false);
-    addFeeEstimates(feeByBlockTarget, estimates);
-  }
-
-   // Add the esplora fee estimates.
-  if (esploraFeeEstimates) {
-    const estimates = processEstimates(esploraFeeEstimates, true, true);
-    addFeeEstimates(feeByBlockTarget, estimates);
-  }
-
-  // Get the minimum fee. If the mempool fee estimates are not available, use a default value of FEE_MINIMUM sat/vbyte as a safety net.
-  const minFee = (mempoolFeeEstimates?.minimumFee ?? FEE_MINIMUM) * 1000;
-  logger.info({ message: 'Using minimum fee: {minFee}', minFee });
-
-  // Filter the estimates to remove any that are lower than the desired minimum fee.
-  feeByBlockTarget = filterEstimates(feeByBlockTarget, minFee);
-
-  return feeByBlockTarget;
-}
+log.info(config.util.toObject());
 
 // Define the layout components.
 
@@ -497,7 +70,7 @@ const Content = (props: { siteData: SiteData; estimates: Estimates }) => (
 
 // Initialize the Express app.
 const app = new Hono();
-logger.info(`Fee Estimates available at ${BASE_URL}/v1/fee-estimates`);
+log.info(`Fee Estimates available at ${BASE_URL}/v1/fee-estimates`);
 
 // Add a health/ready endpoint.
 app.get('/health/ready', async (c) => {
@@ -530,7 +103,7 @@ app.get('/', async (c) => {
     c.res.headers.set('Cache-Control', `public, max-age=${CACHE_STDTTL}`)
 
   } catch (error) {
-    logger.error(error);
+    log.error(error);
     estimates = {
       current_block_hash: null,
       fee_by_block_target: {}
@@ -562,7 +135,7 @@ app.get('/v1/fee-estimates', async (c) => {
     return c.json(estimates);
 
   } catch (error) {
-    logger.error(error);
+    log.error(error);
     return c.text('Error fetching fee estimates', 500);
   }
 });
@@ -573,6 +146,6 @@ export default {
 }
 
 process.on('SIGINT', function() {
-  logger.info({ message: "Caught interrupt signal. Exiting." });
+  log.info({ message: "Caught interrupt signal. Exiting." });
   process.exit();
 });
